Skeleton of a CollectionHelper
The name here is intentionally a bit bad, because naming it feels hard.

This helper class is intended to wrap all of the business logic for
future features such as conditional questions/routing, injecting
contextual/dynamic information, etc. This should make it easy and safe
to reuse the logic across the platform without having to reimplement it
in multiple places.

Most pages that interact with collections/sections/forms/questions
should go through this (and perhaps some additional helper classes we
add later) instead of using the DB models directly.