Lint imports of app.common.data.interfaces.temporary
I've added this module so that we can still write interfaces properly
(to avoid importing DB models), but where we don't expect them to serve
a user-facing purpose. In effect, they only exist while we're
scaffolding the developer-focused pages and when we start applying
domain layers on top, we should be able to drop this module altogether.