Fix intellisense/type hinting for question+submit fields
With the return type of build_question_form just set to FlaskForm,
intellisense+typing doesn't recognise that the `question` and `submit`
fields will exist on the form.

This does a bit of a hack and duplicates the form definition publicly
and privately:

The public (top-level) form definition is used for type hinting.
The private (inside `build_question_form) form definition is used for
instantiating the actual forms and ensuring that multiple requests,
potentially for different questions, all get their own isolated version
of a form. This could be pulled out at some point and cached if we ever
need to optimise.