Add fixture for tracking SQL queries
This is heavily lifted from the implementation of this from
Flask-SQLAlchemy, but we only use Flask-SQLAlchemy-Lite for most of our
DB work now, so we don't have the get_recorded_queries behaviour.

FSL doesn't want to introduce this, so we'll need to provide it
ourselves.

With a small tweak to one of the tests as a demonstration of how to use
it