FSPT-509 - Final tweaks to ensure alignment with designs
These tweaks include:
- Changing the error message for the GGIS Yes / No radio buttons from 'Select yes if you have a GGIS reference number', which felt incorrect, to the inoffensive placeholder 'Please select an option' (we can seek content design input on this)
- Adding the missing maxwords parameter to the description field and using a class constant to ensure a single source of truth for this value
- Changing the caption 'Add a grant' to 'Set up the grant' as per updated designs
- Using formGroup.beforeInput or formGroup.beforeInputs where possible to render paragraph text above form inputs, instead of including them via the hint attribute, which is incorrect and made the text grey
Please note a couple of weirdnesses. Firstly, there is a typo in the govuk_frontend_jinja library for the GovUkCharacterCount widget that means beforeInput is only shown if fromGroup (not formGroup) is true. This needs fixing in the library but in the meantime we can just add the fromGroup parameter to satisfy this condition.
Secondly, we cannot use the formGroup.beforeInput strategy in the case of the GovUkTextInput widget because it renders text to the left of the input box, not above it as per the design. We get around this by using the hint attribute and a custom CSS class color-black (inline styling doesn't work).