FSPT-503: persist dyanmic form data on collection
Add an collections interface method to update the `data` attribute and
store the latest submitted answer from the user.

Note that for now we're just copying the existing data to persuade the
ORM to issue an update on the model, in the future we might want to look
into the SQLAlchemy mutable extension.

This proposes using Pydantic to manage the question data types and give
them a consistent user interface. All of the question types we have
right now are primitives so these work with the current method of
building dynamic forms with a single field.