FSPT-503: calculate collection helper statuses
Replaces the stub methods for calculating form and collection statuses
with methods to iterate over questions/ forms and set not started/ in
progress/ completed statuses accordingly.

Note this commit intentionally out of scopes any template updates to set
an appropriate macro per status, we'll do that and pass in the status
enums in a separate commit.