Increase test run duration in CI
We've seen some of our "slower" tests start to flake in GitHub Actions.
In CI the tests run a lot slower than locally for us, so we'll add a
very gross override that allows tests to take "twice as long" as we
otherwise let them run locally by default.

If we start hitting up against this some more, we should be happy to
consider whether this check is very valuable at all.

For exceptional circumstances, there's now also a marker that lets you
increase (or decrease) the duration for a specific test.