FSPT-520: list collections page
Adds a page to list collections for a given schema, this is hooked up to
the manage a schema (building a tasklist) page.

For now it just uses the collections that are fetched and back
referenced from the colection schema - I think the ORM should only be
fetching the collection schema once for any number of collections which
is nice but should be validated (otherwise thats a very expensive
query!). We might need to revisit that if we want to limit or paginate
the number of collections we're returning.