FSPT-520: collection list add status and tidy up
Adds a full collection schema load option when just fetching the schema
directly, this is to avoid all the properties on the schema then being
lazy loaded when accessed.

This should be covered similarly to the `get_collection` tests to ensure
it doesn't regress and lazy load many more parts, that commit will follow.

Adds the status tag for collections to the list page.