FSPT-524: add collection metadata table
We've called this metadata here as its how we referred to it while
breaking down the story (the users submitted answers as the data of a
collection and what has happened to it as metadata that derives its
status). Another option here is "Collection events" or the things that
have happened to the collection.

SQLAlchemy does reserve the property "metadata" so a prefix was needed
for this naming scheme, happy to revise.

Untriaged notes to be tidied before merging:

commit message, keep users data and metadata separate
- keep the principle of the schema description happening once and not recreating the shape in instances
- avoid creating a separate column on the collection or multiple 1:M tables tracking things like "Submitted", "reviewed", "changes requested", etc.
- this looks to provide a simple middle ground - we should check in on if its doing that or if its actually getting unreasonable

usage:
- if there's an event relevant to my collection and my form its something happening to the form
- if the event is just relevant to my collection its happening to everything (i.e submitted)

extensions:
- could be extended down to updates to the question granalarity
- could be extended even to track the changes that are happening, i.e store the before or after value
- could be extended by adding a comment justification

For now I think it will just be the helper method consuming this, in the future it might be useful to have a helper
specifically to consume this that knows all of the event keys and can give you context specific answers

If this is done using a joined load I don't think you'd be adding much, if we do a selectin we'll be adding an extra query.
We probably want to be able to filter/ get lists of collections without interrogating their metadata/ events but
getting one should always return them.

Edit: I didn't realise collections actually had statuses in the database - maybe we want to either:
- remove that for now OR go with setting the status when we update the data/ issue an event
- a detail on the submission story or a separate ticket?