FSPT-524: form status checks include looking for the complete event
This updates the helper to check for a completed event before
registering the "Complete" status on a form. This will also mean the
entire collection is only considered complete when every form has the
complete event.

For now the "projecting" or determining the salient status from the list
of events is done by the helper itself (in this case just looking for
the presence of an event). In the future we might want a little helper
to translate from the metadata events that have happend into context
specific values (like has something been completed).

Note: for a short period of time between this PR and the next on there
will be no "Complete" collections as the page which can actually
register this event hasn't been merged!