FSPT-524: add have you completed this section component
Extends the check your answers page with a have you completed this
section form component. This optionally allows the user to complete a
section only after all of the answers have been completed.

For now the business logic is in the helper and the interface trusts
that it should be carrying out the operation - we can move that logic
around if needed.