FSPT-524: handle the value error raised when completing a form
In the rare instance of a race condition with another user updating
values or if you're maliciously modifying the pages HTML, it would be
possible to set the property to instruct a form to be completed even
though the answer integrity wasn't complete.

Although this should never happen we should respond to it in a sensible
way and let the user know what was wrong - note that the component won't
be visible for the user when the form answers are incompelte so the edge
case won't have an input to highlight but will still appropriately show
an error summary message.