BAU: Change login required redirect to SSO
Unauthenticated users were previously redirected to the magic link
sign in page and flow. Now that we've identified that our initial
Deliver Grant Funding users will all be logging in with SSO, and now
none of our routes or actions can be taken by non-MHCLG users logging in
via magic link, it makes sense to switch which sign-in flow we push
unauthenticated users to when they first try to login or hit a route
that reqiures them to be authenticated and logged in.