BAU: Add seed-grants for PullPreviews
We recently added a `flask developers seed-grants` command that seeds an
example grant with a basic collection to streamline testing & development.

This adds it to the PullPreview dockercompose so that the grant is also
seeded when a PullPreview environment is spun up on a branch, which
should make it easier to check and test functionality in that environment.