Add 'see all grants' and move to the right
This is a bit of a hack to support rendering service navigation slots
that are pushed to the right of the bar.

We can't feed classes into the slot items directly, so have hacked
together support for having one or two items explicitly right aligned.
We shouldn't ever need more than this.