FSPT-438: send notify email on collection submission
This adds a new notify helper to send a confirmation email after a user
has submitted a collection.

The email will be sent to the user that started that collection
initially (is likely to own and be responsible for it) but we may need
to revise this if multiple people can collaboratively work on a
collection.

The confirmation email failing will cause the submission to be rolled back
using the default session behaviour.