Rename core collections tables
https://mhclgdigital.atlassian.net/browse/FSPT-535

CollectionSchema->Collection
Collection->Submission
CollectionMetadata->SubmissionEvent

And cascade this through the entire codebase and (developer) UI designs.

Given this is an extremely widespread refactor, we have accepted that
we will wipe the databases of all environments fully and lose the
history of this change. It does not feel relevant to retain, and with no
environments that currently handle data that we need to keep around
forever, we can simplify the change significantly by taking this hit.

A `collection` for now is the definition of the data that we want the
user to provide.

A `submission` for now is the actual data the user provides, in line
with the collection.

The goal is for the SubmissionEvent table to be an immutable log of what
has happened to submissions; metadata implies a mutable point in time
K:V representation so call them 'event's to represent immutability.

These terms are still up for grabs pending user research.