BAU: remove dynamic back links and separate tasklist and form management
This commit definitely isn't finished and this flow needs some more
attention but this attempts to bring it back in-line with the sketches
so that we can make improvemnets from a known place.

The original goal with these pages were to separate ordering and
"management/ settings" for tasklist sections from the (hypothesis: much
more frequent) management of questions and flow.

This should make the "back" behaviour a lot more consistent.

This also removes the text that was being used for empty states where
just showing the button to add the thing (form/ section/ question)
should work consistently with other list pages.