FSPT-551: Redirect SSO users to original destination
With the Magic Link journey we were capturing the `session["next"]` value
and redirecting users back to that path when they had successfully
authenticated. We hadn't implemented this for SSO so all users would be
sent to the default Grant List view on successful authentication,
regardless of if they were trying to nagivate to a different page on the
service.

This updates that behaviour to bring it inline with magic links.