Run astral's `ty` across codebase
And apply fixes / ignores to checks that it cannot yet work out.

`ty` is still currently in a pre-production version, so we shouldn't
consider dropping `mypy` yet. However, it looks promising and runs a lot
more quickly than `mypy` does. In part this will be because it's less
comprehensive, but fundamentally when it's ready it is still definitely
going to be an order of magnitude faster.

- `return abort`: this is a bit misleading; abort raises an exception,
  but lots of type checkers don't track this through yet with the
  NoReturn/Never type. Eventually we should be able to drop this
  syntax, but for now it helps type checkers realise that any code
  following this will never run.
- mypy ignore unused: ty is picking up some errors that mypy is not;
  this may or may not be correct, but while we're running both type
  checkers we'll need to expect some inconsistency between the two.