FSPT-537: Add a `mode` to submission table
We want to support the idea of separating submissions into test and live
submissions.

Test submissions will help platform and grant admins gain confidence
that their configuration is correct and will work for real users.

Live submissions contain the real data submitted by grant
recipients/applicants.

Test submissions will still work with the end-to-end system, but will
only appear when a platform user is in the context of testing a
grant/collection.

In the future, we may further develop the idea of a 'Preview
submission', that does not appear in the end-to-end system but allows a
subset of grant admins (grant relationship managers and content
designers) to very quickly validate changes to a collection with an
extremely-temporary/transient/ephemeral submission instance.