Add test mode banner to relevant templates
Applies an arbitrary banner/signage to all of the pages that can view
test or live submissions, across both form runner and the developers
tab.

This proves the concept but does not deep dive the design. When building
out the user-facing variants of this, we should pay attention to
architecting the code and templates in a way that the test data banner
is reliably displayed for all test entities, without necessarily having
to do it manually every time.