FSPT-561: Evaluate (python-like) expressions
We're starting to bring in the idea of validation and conditional
routing to the form builder/runner. We need to be able to describe how
these rules work and want a simple, powerful, and flexible system that
will let us do straight-forward and complex evaluations. These might
vary from things like "is the answer bigger than 100", to "if the answer
to question A was X, then this answer must be lower than the sum of all
their previous reported spending".

We feel like it's possible to have a single unified system under the
hood that can handle all of validation, conditional question behaviour,
and injecting values into text (eg questions and error messages) and
think it would be simpler to manage and maintain than doing anything too
unique for each of those features.

The idea here is to allow writing (a subset of) Python statements and
evaluating those. Traditionally this was done with `eval()` but there
are many and varied security concerns with doing this. Using Python-like
language should simplify learning how to write expressions (at least for
developers), but as this will handle untrusted user input we need to be
extremely cautious in our approach so as not to create huge security
vulnerabilities. It absolutely must be impossible for expressions to
result in arbitrary code execution.

We use a library called `simpleeval` which takes an expression and
converts it to the abstract syntax tree (AST) and then manually
processes that tree. We can easily add or remove nodes that we want to
be allowable in expressions, which should provide a good level of
protection against dangerous operations, as long as we configure the
allowlist of nodes correctly.

With this system, an expression might look like any of these:

`answer < 100`
`answer == "hello"`

In the future we might be able to do more interesting things like:
`answer <= grant.allocated_funding`
`answer <= question_a.answer`
`answer <= sum(forecasted_spends)`

With this expression system we're able to define the context available
to the expression (eg the set of variables/functions that can be used).
This allows us to expose some data/context to the expression, such as
the question's answer, or answers to all other questions, or information
about the grant they're providing data for.

But a lot of this is left as an extension exercise for the future.