Change UserRole constraint
We have agreed that we will treat roles as strictly hierarchical for
now, ie ADMIN is a superset of the permissions granted by MEMBER. Based
on this it makes sense to only allow one UserRole entry to exist for the
combination of user+org(nullable)+grant(nullable). So let's enforce
that.

We may switch away from single-role-based to fine-grained permissions or
multi-role in the future, but for now let's enforce our current
approach.