FSPT-534 - Delegating case-sensitive checks to database
When we add or change a grant, we need to do a case-insensitive check to see if the grant name exists in the database. Currently we implement that check in the Python which means manually lowercasing the strings before comparison, and failing to leverage the CITEXT PostgreSQL datatype that handles case-insensitive checking for us. The same is true for user email address when we add new users to grants.

This change removes the Python lowercasing logic and delegates responsibility for the check to the database, where case insensitivity is handled out of the box, with refactoring.