FSPT-564: move managed expression tests
Organises managed expression tests so they reflect the file structure.

Add an explicit scenario to make sure we're covering questions being
filtered out of "supported" questions to depend on for a condition.

Don't handle the value error thrown when parsing a managed expression
form - there should be no way of triggering this without actively
manipulating the frontend or attempting to circumvent the default logic.
Instead of showing the user a message they can't do anything about we'll
raise this as an exception which will propegate to Sentry.