FSPT-550: Make SSO Stub Server user azure ad subject id
Now we're using the unique azure_ad_subject_id to uniquely identify users
in our database on authentication, we need to update the stub server to
match this behaviour without impacting its usability.

This change creates a hash of the email address used to login via the
stub server as the unique azure_ad_subject_id returned by the stub server
so that we can consistently log in as the same user without having to
worry about generating this value.