FSPT-594 Create managed validation rules
Allow the creation of "managed" validation rules against questions.
"managed validation" here means that we take care of building the
validation expression for the user and can therefore provide them a very
simple interface to add validation.

For now, the only managed validation we can reasonably support is
numeric validation against an integer field.

The first example bit of validation is "the answer must be greater than
x". We'll extend this in the future.