FSPT-594: Support inclusive integer validation
Allow GreaterThan, LessThan, and Between validation to include or
exclude the provided value. Essentially this lets us support gt, gte,
lt, lte, x < answer < y, x <= answer <= y (and variants).

The UI design of this is questionable and we may want to split this out
into explicit radios, but for now this demonstrates the principle.

We may also want to split this out into separate managed expressions,
but again for now this demonstrates the principle in an acceptable way.