FSPT-594: limit validation to one instance of each type
It doesn't make sense to have two GreaterThan validations for the same
question with different configuration. This patch creates a partial
index on the expressions table that enforces a maximum of one of each
kind of validation on a specific question. You could have a GreaterThan
validation and a LessThan validation on one question, but not two
GreaterThan validations.

Handling this in a user-friendly way is left to the UI.