FSPT-594: switch includes to macros
These more cleanly encapsulate the requirements (arguments) needed to
render the form. I think includes are fine for repetitive text where you
don't rely on arbitrary context from the parent template, or where
you're _really_ being strained on the limits of duplication.

That's not the case here, so it feels cleaner to call a macro and pass
it the things you need.