FSPT-594: Handle error on un-validatable question types
Rather than raising an error when trying to access the 'add validation'
page for a question that has no viable managed validators, we can render
the page and show content that tells the user this.

It's not the world's best user experience, but it's better than a 500
error, and we can refine the design when we deal with the situation for
real.