BAU: Unpack user upsert values() arguments
Previously for our upsert user behaviour we were passing in every field
to the `values()` regardless of it was set or not, leading to incorrect
values entering the database (eg. `token` for the user's `name`).

Instead we can unpack the `on_conflict_set` dictionary and only pass in
the values that have been set.