Bugfix: Allow grant team members to sign in via SSO on first sign-in
The recent implementation of identifying SSO users via their
azure_ad_subject_id missed an important case - grant team members who
have not yet logged in will not have an azure_ad_subject_id, but will have
a user record created. This means when they try to login via SSO, the
flow will break as they won't be found as an existing user by
get_user_by_azure_ad_subject_id, and they won't get upserted because
a user with that email already exists but the unique lookup is on
azure_ad_subject_id so the upsert is failing silently.

This slight hacky fix falls back to getting the user by email and
setting the azure_ad_subject_id when they first log in, allowing the rest
of the sso_get_token flow to happen as previously set.

This will be fixed and should be simplified in FSPT-515 which will add
an 'invitation' mechanism for grant team members and therefore not
directly create a User record for them until they first sign in.