FSPT-595: summary list layout for validation
Use a summary list for displaying validations, which is likely to be
similar to the layout for conditions.

Unlike conditions, it's slightly less clear (to me at least) what a good
structure here would be for the key/value columns. However, also unlike
with conditions, the ordering of validation may matter - they do run
sequentially and don't all need to pass in combination. Therefore it
feels important to represent the order in some way. So here we use the
ordering as the key and the value as the name/description of the
validation (and the link to change it).