Streamline dev experience for new managed expressions
This changeset will be rough to review; the best review is probably
looking wholesale at the actual end-result, which sits across three
files:
- app.common.expressions.forms
- app.common.expressions.managed
- app.common.expressions.registry

--------

A bit of an exploration for improving the DX/reducing the number of
touchpoints that need to be edited when adding a new managed expression.

Before this change, adding a new managed expression meant changes across
the following set of places:
- app.common.data.types
  Update `ManagedExpressionsEnum` to define the name of the new
  expression.

- app.common.expressions.forms
  Define a form for the question type that explicitly lists all of the
  managed expression types it supports.

- app.common.expressions.helpers
  Add the above form to a lookup table mapping it against a question
  data type.

- app.common.expressions.managed
  Define the new managed expression as a pydantic model; this describes
  the data needed to define an instance.

- _number_expression.html
  Render the fields for the new expression into a variable and pass them
  into the render for `form.type`. You need to manually care about the
  order you're passing the blobs of HTML into the `type` field
  rendering, because they're implicitly linked.

- generating a migration for the managed expression enum change.

--------

And then if you're also adding a new question type altogether, you need
to:

- add_question_condition_select_condition_type.html
  Check the question type and render the equivalent of
  _number_condition.html

- add_question_validation.html
  Check the question type and render the equivalent of
  _number_validation.html

- app.common.data.migrations.versions (new migration for
  the managed expression enum change).

--------

This patch reduces the number of touchpoints down to two, in theory:
- app.common.data.types (add an entry to the ManagedExpressionsEnum)
- app.common.expressions.managed (create the new managed expression by
extending the BaseExpression class).

Each managed expression is now a slightly thicker class that declares
both the data needed to define it (the pydantic model equivalent), and
the form fields+metadata needed to collect that successfully from a form
builder user. This mixes purposes slightly, but does co-locate all of
the changes in one place which - theoretically - should make it a
simpler task. At least that's the theory I guess I was exploring here.