FSPT-608: combine add+edit validation templates
These two templates are extremely similar, with (currently) just
different text for the heading and submit button. So let's combine them.

It is an open question as to whether or not the 'edit validation' page
makes complete sense being the same, or if only the fields for that
specific managed validation type (eg greater than) should be visible
(instead of the full list of all managed validation available).

However I think this is fine for now, and we can tweak with full design
or feedback.