FSPT-619: delete validation rules
Following the deletion pattern laid out for collections, sections,
forms, and questions, let's also make it possible to delete validation
rules from the "edit validation" page.

We may be able to optimise this (eg remove the warning step) in the
future as the decision is very lightweight and should be easy to
reverse, but this feels fine for now. Design should be able to tackle
this as part of the domain work.