FSPT-550: Expire user object on role create/delete
When creating or deleting roles, the linked `user` object is not
updated by SQLAlchemy until the end of the transaction. This is fine for
simple transactions but when the route or function using the db interfaces
needs to reference the updated `user` object this causes a problem as
it will not have refreshed to reflect the updated relationship to the
UserRole table.

To avoid this, we're explicitly adding `db.session.flush()` and
`db.session.expire(user)` to interfaces that create and delete user roles
to automatically expire the passed `user` object. If the route/function
then tries to reaccess the user object it will query the database to get
the updated object again. By using `expire` rather than `refresh` we only
do this subsequent query if the object is accessed again, otherwise no
additional queries are made.

This also updates the `user` interfaces to take the whole user object as
opposed to just the UUID.