FSPT-564: list conditions for a given question
This commit adds the list of conditions that have been assinged to a
question on the edit question page. It's able to use some of the core
helpers that were included in listing validations.

Extends the base ManagedExpression class to let the template lookup the
question being referenced in the context of a managed condition.

Note this doesn't currently let you edit conditions - this is because
how conditions are added is slightly out of date with validations which
has set a pattern for this - there will be a follow up PR that brings adding
managed condition "forms" in-line with validations which should allow us
to edit conditions in a very similar pattern (and benefts from using all
of the same managed expressions as for validations consistently).