FSPT-564: replace the specific conditions form with expression form
It's likely that for a given question type we'll be sharing some if not
all managed expressions (simple ways of comparing that data) between
conditions and validations.

This combines the two forms for adding an integer expression which were
put together at different times into one - this should mean devs can
specify one clear interface for managed expressions for each question
type and that will consistently apply to conditions and validations.