FSPT-564: add expression unique constraint for conditions
This adds a unique constraint for conditions on a question, you should
only be able to add one condition per question and comparison.

This makes the decision to allow you to depend on one question with
different comparisons (i.e show up if question A is greater than 1000,
separately show up if question is A is less than 2000) because that
makes the behaviour consistent with validation uniqueness. We can always
tweak this later if we think its only sensible to have one managed
expression per dependant question.