FSPT-564: handle duplicate errors for conditions
Updates coverage and adds handlers when trying to add or update a
condition that runs into a duplicate value error.

The same disclaimer applies as for validation - we'll probably want a
slicker way of handling this error for the user (maybe linking to the
condition that you're trying to duplicate so you can go and update that
instead or something) but this should be all we need to ensure our
configuration integrity and we can update this if users have issues.