FSPT-629: move the managed expression key to the databse model
This proposes moving the "key" out of the JSON context for managed
expressions and associating it with the database row. This feels more
appropriate as the type of managed expression isn't relevant to the
expression statement context (when its evaluated, the context is
passed in) and its something that we check when deserialising
expressions.

This commit out of scopes thinking about how this would work with
existing data.