BAU: remove the empty grant home page
Remove the empty grant home page until we have a use for it - it
currently makes the build look broken where its actually functional and
should encourage people to critisise about it as a functional piece of
software.

As soon as we have information to put on a dashboard/ homepage (overview
of your currently open rounds? stats on your open applications/
assessments? number of reports on funding progress you have left to sign
off? etc.) we can add this back in.