BAU: fix the grant details page
Spotted while making this the default page instead of the home page. Any
time we use the govuk grid system rows should have columns, even when
its full width (it deals with margins etc.).

This will line the page up the service nav and all the other pages
(as-is is slightly out of the margin at the left).