BAU: tidy up the grants team member page
Happy for this to be closed or reverted if its coming in future tickets.

The primary change here is to remove the "Contact us" prompt if the user
already has permissions to add a new user - this should only need to be
shown to users that don't have permissions to do it themselves (the page
currently limits this to users with administrative permissions).

When thats removed the page looks empty and broken without any users -
bring in the empty no users disclaimers from the prototype to avoid
this.

Misc.
- moves xl headings to l heading as we've agreed we'll use those
  consistently now
- use empty lists being falsey for what feels like a more readable
  template to me
- the SQLAlchemy model for Grant will always populate an empty list of
  users, no need to guard against it being missed for readability